refactor: rename directories from singular to plural
