+++ src/green-point/green-point.bootstrap.ts
@@ -1,6 +1,6 @@
 import { type FastifyInstance } from 'fastify'
 import type GreenPointRepository from './domain/repositories/green-point.repository'
-import GreenPointHandler from './infrastructure/handler/green-point.handler'
+import GreenPointHandler from './infrastructure/handlers/green-point.handler'
 import GreenPointMikroORMRepository from './infrastructure/repositories/mikro-orm/green-point.mikroorm.repository'
 import GreenPointRoute from './infrastructure/routes/green-point.route'
