+++ src/neighbor/infrastructure/swagger-schemas/neighbor.swagger-schema.ts
