+++ src/responsible/infrastructure/swagger-schemas/responsible.swagger-schema.ts
