+++ src/reward-partner/infrastructure/swagger-schemas/reward-partner.swagger-schema.ts
