+++ src/waste-category/infrastructure/handlers/waste-category.handler.ts
